core metadata
